GraphQL::ObjectType was deprecated and removed
The GraphQL::ObjectType was deprecated and removed in graphql 2.0.0.
When this project was upgraded to 2.0.0 the code failed because it
was checking if types were GraphQL::ObjectType.

By checking if GraphQL::ObjectType is defined before checking if the
type is_a GraphQL::ObjectType the library should be compatible across
multiple upstream versions.