fix: camelize string fields to match sym behavior
Followup to a previous feature release, which intentionally did not
camelize string inputs in order to preserve backwards compatibility.
Here we are breaking compatibility to make the behavior consistent.
Users can revert the behavior using the `camelize` option if it is
necessary.

BREAKING CHANGE: string fields will be camelized by default rather than
passed as is.