feat: Support Federation v2 (#196)
* Add support for the @shareable directive

* Fix rubocop offense

* Add support for the @inaccessible directive

* Add support for the @override directive

* Refactor to appease rubocop

* Replace `Hash#except` with `Hash#delete_if`

Oops! I didn't realise `Hash#except` was only available from Ruby 3.0

* Update README to include v2.0 directives

* Fix incorrect handling of keyword args in field

* Import federation directives into subgraph to opt into federation 2 (hack!)

See: https://www.apollographql.com/docs/federation/federation-2/moving-to-federation-2/#opt-in-to-federation-2
This is a horrible hack. Need to figure out if there is a clean way of including this in the document and the printer output.

* Enable federation 2 via a `federation_2` class method on Schema

* Change `federation_2` to `federation version: 2`

This is more explicit and extendable in the future if we ever need any other federation options at the schema level

* Add a unit test for Schema.federation_version

* Update README with how to opt in to Federation v2

* Update federation version to support semantic versioning

* Add `federation__` namespace prefix to all directives when using federation version >= 2.0

* Refactor the `merge_directives` method

This approach feel a little easier to understand

* Fix grammar issues in test names