feat: adds support for Interface Entities
Context: In Federation v2.3 there was added support for interfaces to be
Entities. There are two ways to consume this new feature.

1. Extend an Interface Entity with the `@interfaceObject` directive,
   when the Entity is defined by a different subgraph.

2. Implement the Interface Entity with its implementing Type Entitites.

There was previos work into getting the first part covered in this gem,
but we were still lacking support for the 2nd part. This work address
that.

The main issue here is the fact that an `Union` in `GraphQL` can't be an
`Interface` according to the [spec](https://spec.graphql.org/October2021/#sec-Unions.Type-Validation), but at the same time, according to the Apollo Federation [spec](https://www.apollographql.com/docs/federation/federated-types/interfaces), an interface can be an Entity, and an Entity is an Union. Therefore, we have to extend the validation (`assert_valid_union_member`) for the Entity union to allow `Interfaces` (more exactly `Modules`) as possible types.