fix: call `resolve_references` once per `__typename`
The previous implementation called `resolve_references` once per
contiguous `__typename` to ensure the correct ordering of the response.
This could have resulted in wildly varying runtime efficiencies.
A request calling for N unique types across M entities could have
as little as N queries or as many as M queries, depending on if the
entities happened to be in large groups of the same type or very
interleaved.

By keeping track of the requested representations and their original
indices, we can rebuild the response in the correct order even when
things get shuffled during computation.