chore: update rubocop
jaro_winkler, a dependency of the ancient rubocop version, no longer
compiles on macOS and modern Ruby.  so update rubocop. err on the side
of not enabling new cops. a few minor new offences were corrected,
a few ignored inline with rubocop:todo, and a couple that had several
instances were ignored in the config