Cleanup gemfiles.lock files and stabilize CI
It's a bad practice to commit Gemfile.lock files into gem repositories - `gemspec` files should be the source of truth for dependencies.
I removed all Gemfile.lock files and updated the CircleCI configuration to use the Gemfile for cache keys.

Additionally, I updated the `rack` development dependency in the gemspec and locked it to version 2.0 because of the breaking changes in version 3.